Apply newly-discovered symbols (#117)
* Fix ulong in ut_Color.h
* Correct computation for `dEn_c::smc_DEADFALL_GRAVITY`
* Rename unused structures in `fBase_c`

---------

+++ include/game/bases/d_res.hpp
@@ -25,7 +25,7 @@ public:
          * @param path The path to the current resource.
          * @return The resource data.
          */
-        virtual void *execute(void *data, u32 folderSig, const char* path) = 0;
+        virtual void *execute(void *data, ulong folderSig, const char* path) = 0;
     };
 
     /// @brief A callback class for EGG::Archive::searchInside.
