+++ include/game/bases/d_res_mng.hpp
@@ -13,7 +13,7 @@ private:
     class resCallback_c : public dRes_c::callback_c {
     public:
         virtual void init(const char *name);
-        virtual void *execute(void *data, u32 folderSig, const char* path);
+        virtual void *execute(void *data, ulong folderSig, const char* path);
     };
 
 public:
